requirements.txt
Flask==2.3.3
requests==2.31.0
PyYAML==6.0.1
schedule==1.2.0
matplotlib==3.8.0
seaborn==0.12.2